requirements.txt
# Core dependencies - using compatible versions
Flask>=3.0.0,<4.0.0
langchain-perplexity>=0.1.0
langchain-core>=0.3.0
pydantic>=2.0.0,<3.0.0

# Utilities
python-dotenv>=1.0.0

# Production
gunicorn>=20.1.0